refactor: re-organise modules
