+++ src/composables/deepUnref.ts
@@ -1,4 +1,4 @@
-import utilities from "@/core/utilities";
+import utilities from "@/core/utils/functions";
 
 export function deepUnref<T>(arg: T): T {
 	return utilities.deepCopy(arg);
