--- src/core/saves.ts
@@ -1,237 +1,0 @@
-/* eslint-disable @typescript-eslint/no-explicit-any */
-import Ajv from "ajv";
-import schemaLegacy from "@/schemas/save_format_legacy.json";
-import schemaV1 from "@/schemas/save_format_v1.json";
-import schemaV2 from "@/schemas/save_format_v2.json";
-import schemaV3 from "@/schemas/save_format_v3.json";
-import { SaveFormat, SerializedState } from "@/store";
-import { CardCategory, createRootFolder } from "./model/cards";
-import utilities from "./utilities";
-
-export enum SaveVersion {
-	Legacy = "save-legacy",
-	v1 = "save-v1",
-	v2 = "save-v2",
-	v3 = "save-v3",
-	Latest = "save-v3",
-}
-
-const validator = new Ajv()
-	.addSchema(schemaLegacy, SaveVersion.Legacy)
-	.addSchema(schemaV1, SaveVersion.v1)
-	.addSchema(schemaV2, SaveVersion.v2)
-	.addSchema(schemaV3, SaveVersion.v3);
-
-abstract class SaveProcessor {
-	inputSaveVersion: SaveVersion;
-	targetSaveVersion: SaveVersion;
-
-	constructor(inputSaveVersion: SaveVersion, targetSaveVersion: SaveVersion) {
-		this.inputSaveVersion = inputSaveVersion;
-		this.targetSaveVersion = targetSaveVersion;
-	}
-
-	/**
-	 * Validate data from a save file against a JSON schema corresponding to the specified save format.
-	 * @param save the save data to validate
-	 * @param format the save format descriptor. Describes the schema against which data will be validated. Will throw an error if format does not describe a registered schema.
-	 * @returns a boolean indicating if the data validates against the specified save format.
-	 */
-	static validate(save: SaveFormat | object, format: SaveVersion) {
-		const isValid = validator.validate(format, save);
-		if (!isValid && process.env.NODE_ENV !== "production") console.error(validator.errors);
-		return isValid;
-	}
-
-	validateAndConvert(save: any): any {
-		if (SaveProcessor.validate(save, this.inputSaveVersion)) return this.convert(save);
-		else
-			throw new Error(
-				"Save format does not match specified save version: save data is unusable."
-			);
-	}
-
-	protected abstract convert(save: any): any;
-}
-
-class V1SaveProcessor extends SaveProcessor {
-	constructor() {
-		super(SaveVersion.Legacy, SaveVersion.v1);
-	}
-
-	/**
-	 * Convert legacy-format data to v1-format.
-	 * Changes between legacy and v1 formats are:
-	 * - New field: '_category' in each card
-	 * - New field: '_meta' on root object
-	 * - New field: 'quickNote' on root object
-	 * - Changes to Objectives
-	 * - New Faction category
-	 * - Remove plural on card category array fields
-	 *
-	 *
-	 * @param save data to convert to v1 format
-	 * @returns input data converted to v1 format
-	 */
-	protected convert(save: any): any {
-		// Convert root object:
-		// - Add '_meta' property
-		// - Add 'quickNote' property
-		// - Keep previous fields
-		const converted = {
-			_meta: {
-				version: SaveVersion.v1,
-				lastUpdate: new Date().toISOString(),
-			},
-			name: save.name,
-			days: save.days,
-			season: save.season,
-			cards: {} as any,
-			quickNote: "",
-		};
-
-		// Convert cards:
-		// - Transform Objectives to Quests
-		// - Use category constant as field name (i.e. remove plural on card fields)
-		// - Add corresponding '_category' field on each card
-		// - Add new Faction category
-		converted.cards.quest = save.cards.objectives.map((entry: any) => {
-			return {
-				_category: "quest",
-				id: entry.id,
-				tags: entry.tags,
-				title: entry.desc,
-				tasks: [],
-			};
-		});
-
-		converted.cards.event = save.cards.events.map((entry: any) => ({
-			_category: "event",
-			...entry,
-		}));
-		converted.cards.location = save.cards.locations.map((entry: any) => ({
-			_category: "location",
-			...entry,
-		}));
-		converted.cards.character = save.cards.characters.map((entry: any) => ({
-			_category: "character",
-			...entry,
-		}));
-		converted.cards.note = save.cards.notes.map((entry: any) => ({
-			_category: "note",
-			...entry,
-		}));
-		converted.cards.faction = [];
-
-		return converted;
-	}
-}
-
-class V2SaveProcessor extends SaveProcessor {
-	constructor() {
-		super(SaveVersion.v1, SaveVersion.v2);
-	}
-
-	/**
-	 * Convert v1-format data to v2-format.
-	 * Changes between v1 and v2 formats are:
-	 * - Convert each card list to a file tree
-	 *
-	 * @param save data to convert to v2 format
-	 * @returns input data converted to v2 format
-	 */
-	protected convert(save: any): any {
-		const converted = utilities.deepCopy(save);
-
-		// For each category, create a root folder and set all existing cards as its files
-		Object.keys(converted.cards).forEach((category) => {
-			const folder = createRootFolder(category as CardCategory, converted.cards[category]);
-			converted.cards[category] = folder.serialize();
-		});
-
-		return converted;
-	}
-}
-
-class V3SaveProcessor extends SaveProcessor {
-	constructor() {
-		super(SaveVersion.v2, SaveVersion.v3);
-	}
-
-	/**
-	 * Convert v2-format data to v3-format.
-	 * Changes between v2 and v3 formats are:
-	 * - Add a 'tags' list to the folders' metadata
-	 * - Add a 'desc' fields to quests
-	 *
-	 * @param save data to convert to v3 format
-	 * @returns input data converted to v3 format
-	 */
-	protected convert(save: any): SerializedState {
-		const converted = utilities.deepCopy(save);
-
-		// Add a new 'tags' array in every folder's metadata
-		Object.keys(converted.cards).forEach((category) => {
-			const rootFolder = converted.cards[category];
-			Object.keys(rootFolder).forEach((folderID) => {
-				rootFolder[folderID].metadata.tags = [];
-			});
-		});
-
-		// Add a new 'desc' field on every quest
-		Object.keys(converted.cards.quest).forEach((folderID) => {
-			const folder = converted.cards.quest[folderID];
-			folder.files.forEach((quest: any) => {
-				quest.desc = "";
-			});
-		});
-
-		return converted;
-	}
-}
-
-function getProcessorFromInputVersion(inputSaveVersion: SaveVersion): SaveProcessor | undefined {
-	switch (inputSaveVersion) {
-		case SaveVersion.Legacy:
-			return new V1SaveProcessor();
-		case SaveVersion.v1:
-			return new V2SaveProcessor();
-		case SaveVersion.v2:
-			return new V3SaveProcessor();
-		case SaveVersion.v3:
-		case SaveVersion.Latest:
-			return undefined;
-	}
-}
-
-function buildConversionPipeline(inputSaveVersion: SaveVersion) {
-	const toPipe = [];
-	let nextInput = inputSaveVersion;
-	while (nextInput !== SaveVersion.Latest) {
-		const proc = getProcessorFromInputVersion(nextInput);
-		if (proc) {
-			nextInput = proc.targetSaveVersion;
-			toPipe.push((arg: any) => proc.validateAndConvert(arg));
-		} else break;
-	}
-
-	return utilities.pipe(...toPipe);
-}
-
-export default {
-	/**
-	 * Convert save data to the latest format. If the save is already at the latest version, it is returned as is.
-	 * This method throws an error if the format of the specified save cannot be identified and as such cannot be converted.
-	 *
-	 * @param save the save data to convert
-	 * @returns the save data, converted if necessary to the latest save format.
-	 */
-	ensureLatestVersion(save: any): SaveFormat {
-		const inputSaveVersion = (save?._meta?.version as SaveVersion) ?? SaveVersion.Legacy;
-		if (Object.values(SaveVersion).includes(inputSaveVersion)) {
-			const pipeline = buildConversionPipeline(inputSaveVersion);
-			const saveAtLatestFormat: SaveFormat = pipeline(save);
-			return saveAtLatestFormat;
-		} else throw new Error("Save format could not be identified: save data is unusable.");
-	},
-};
