+++ src/core/translation/index.ts
@@ -1,0 +1,1 @@
+export * from "./translation";
