+++ src/core/utils/functions.ts
@@ -1,11 +1,5 @@
 import { klona } from "klona/lite";
 
-// Type utilities
-export type UnionOfTupleValues<T extends Readonly<string[] | number[] | symbol[]>> = T[number];
-export type Constructor<T> = new (...args: any[]) => T;
-
-// Map utility functions
-
 /**
  * Merge mapB onto mapA.
  *
