+++ src/core/utils/types.ts
@@ -1,0 +1,3 @@
+export type UnionOfTupleValues<T extends Readonly<string[] | number[] | symbol[]>> = T[number];
+
+export type Constructor<T> = new (...args: any[]) => T;
