+++ src/store/index.ts
@@ -4,8 +4,8 @@ import { computed, reactive } from "vue";
 import { deepUnref } from "@/composables/deepUnref";
 import { LocalStorageKey, Season } from "@/core/constants";
 import { CardsStoreSerialized } from "@/core/model/cards";
-import saves, { SaveVersion } from "@/core/saves";
-import utilities from "@/core/utilities";
+import saves, { SaveVersion } from "@/core/persistence/save-converter";
+import utilities from "@/core/utils/functions";
 import { useCampaignInfoStore } from "./campaignInfo";
 import { useCardsStore } from "./cards";
 import { useQuickNoteStore } from "./quickNote";
