+++ src/store/preferences.ts
@@ -2,7 +2,7 @@ import { defineStore } from "pinia";
 import { ref, watch } from "vue";
 import { useTheme } from "vuetify";
 import { LocalStorageKey } from "@/core/constants";
-import { SupportedLanguages } from "@/core/translation";
+import { SupportedLanguages } from "@/core/translation/translation";
 import type { Theme } from "@/plugins/vuetify";
 
 export type { Theme } from "@/plugins/vuetify";
